chore: open gov component refactoring (#909)
