+++ packages/renderer/src/renderer/screens/OpenGov/Overview/types.ts
@@ -1,0 +1,12 @@
+// Copyright 2024 @polkadot-live/polkadot-live-app authors & contributors
+// SPDX-License-Identifier: GPL-3.0-only
+
+export interface OverviewProps {
+  setSection: React.Dispatch<React.SetStateAction<number>>;
+  setSectionContent: React.Dispatch<React.SetStateAction<string>>;
+}
+
+export interface OverviewExploreProps {
+  setSection: React.Dispatch<React.SetStateAction<number>>;
+  setSectionContent: React.Dispatch<React.SetStateAction<string>>;
+}
